Refactor test locations.
