+++ cypress/fixtures/preApprovals/requests/createRobustAccessorial.js
@@ -2,7 +2,7 @@ export function createItemRequest({ shipmentId, csrfToken, code, quantity1 }) {
   const tariff400ng = {
     '105B': { item: 'Pack Reg Crate', id: 'deb28967-d52c-4f04-8a0b-a264c9d80457', location: 'ORIGIN' },
     '105E': { item: 'UnPack Reg Crate', id: '6df4f1aa-a232-4eef-bbe8-f06bfb0b6d40', location: 'DESTINATION' },
-    '35A': { item: 'Third Party Service', id: '6df4f1aa-a232-4eef-bbe8-f06bfb0b6d40', location: 'EITHER' },
+    '35A': { item: 'Third Party Service', id: 'c6a865dd-324a-48a5-9b03-5db8dcd044d1', location: 'EITHER' },
   };
   let itemDetails;
   if (code in tariff400ng) {
