+++ src/shared/preApprovals.js
@@ -5,6 +5,9 @@ export const isNewAccessorial = item => {
   if ((code === '105B' || code === '105E') && !item.crate_dimensions) {
     return false;
   }
+  if (code === '35A' && !item.estimate_amount_cents) {
+    return false;
+  }
 
   return true;
 };
